Print the full log output of an admixr wrapper to the console.

Print the full log output of an admixr wrapper to the console.

loginfo(x, target = NA, save = FALSE, prefix = NA, dir = ".", suffix = ".txt")

Arguments

x
Output from one of the admixr wrappers (d, f4, qpAdm, ...)

target
A specific log to examine (relevant for multiple target qpAdm runs)

save
Save the log output to a disk?

prefix
Prefix of the output log file(s) (name of the admixr command by default)

dir
In which directory to save the log file(s)?

suffix
Suffix of the output log file(s) (".txt" by default)

Examples

# download an example genomic data set and prepare it for analysis
snps <- eigenstrat(download_data(dirname = tempdir()))

# define a set of populations to analyze and calculate a D statistic
pops <- c("French", "Sardinian", "Han", "Papuan", "Khomani_San", "Mbuti", "Dinka")
result_d <- d(
    W = pops, X = "Yoruba", Y = "Vindija", Z = "Chimp",
    data = snps
)

# examine the full log output associated with the returned object
loginfo(result_d)

#> Full output log of qpDstat:
#> ===================================================
#> 
#> qpDstat: parameter file: /var/folders/t7/9gjtb6m92flbnp930618vt3r0000gn/T//Rtmpv2F4wc/qpDstat__15890178.par
#> ### THE INPUT PARAMETERS
#> ##PARAMETER NAME: VALUE
#> genotypename: /var/folders/t7/9gjtb6m92flbnp930618vt3r0000gn/T//Rtmpv2F4wc/snps/snps.geno
#> snpname: /var/folders/t7/9gjtb6m92flbnp930618vt3r0000gn/T//Rtmpv2F4wc/snps/snps.snp
#> indivname: /var/folders/t7/9gjtb6m92flbnp930618vt3r0000gn/T//Rtmpv2F4wc/snps/snps.ind
#> popfilename: /var/folders/t7/9gjtb6m92flbnp930618vt3r0000gn/T//Rtmpv2F4wc/qpDstat__15890178.pop
#> printsd: YES
#> ## qpDstat version: 970
#> number of quadruples 7
#>   0               French    1
#>   1            Sardinian    1
#>   2                  Han    1
#>   3               Papuan    1
#>   4          Khomani_San    1
#>   5                Mbuti    1
#>   6                Dinka    1
#>   7               Yoruba    1
#>   8              Vindija    1
#>   9                Chimp    1
#> jackknife block size:     0.050
#> snps: 500000  indivs: 10
#> number of blocks for block jackknife: 547
#> nrows, ncols: 10 500000
#> result:     French     Yoruba    Vindija      Chimp      0.0313      0.006933     4.510   15802  14844 487753 
#> result:  Sardinian     Yoruba    Vindija      Chimp      0.0287      0.006792     4.222   15729  14852 487646 
#> result:        Han     Yoruba    Vindija      Chimp      0.0278      0.006609     4.199   15780  14928 487925 
#> result:     Papuan     Yoruba    Vindija      Chimp      0.0457      0.006571     6.953   16131  14721 487694 
#> result: Khomani_San     Yoruba    Vindija      Chimp      0.0066      0.006292     1.051   16168  15955 487564 
#> result:      Mbuti     Yoruba    Vindija      Chimp     -0.0005      0.006345    -0.074   15751  15766 487642 
#> result:      Dinka     Yoruba    Vindija      Chimp     -0.0009      0.006124    -0.151   15131  15159 487667 
#> ##end of qpDstat:        4.449 seconds cpu        0.000 Mbytes in use