Calculate ancestry proportions in a set of target populations.

Calculate ancestry proportions in a set of target populations.

qpAdm(
  data,
  target,
  sources,
  outgroups,
  outdir = NULL,
  params = list(allsnps = "YES", summary = "YES", details = "YES")
)

Arguments

data
EIGENSTRAT data object.

target
Vector of target populations (evaluated one at a time).

sources
Source populations related to true ancestors.

outgroups
Outgroup populations.

outdir
Where to put all generated files (temporary directory by default).

params
Named list of parameters and their values.

Examples

# \dontrun{
# download example data set and prepare it for analysis
snps <- eigenstrat(download_data())

# estimate the proportion of Neandertal ancestry in a French
# individual and other associated qpAdm statistics (see detailed
# description in the tutorial vignette)
result <- qpAdm(
    target = "French",
    sources = c("Vindija", "Yoruba"),
    outgroups = c("Chimp", "Denisova", "Altai"),
    data = snps
)
# extract only the main table of admixture proportions
result$proportions

#> # A tibble: 1 x 8
#>   target Vindija Yoruba stderr_Vindija stderr_Yoruba nsnps_used nsnps_target
#>   <chr>    <dbl>  <dbl>          <dbl>         <dbl>      <dbl>        <dbl>
#> 1 French   0.022  0.978          0.006         0.006     500000       499434
#> # … with 1 more variable: pvalue <dbl>

# }